User and filter to entities

+++ Application/Notes/Delete.cs
@@ -1,5 +1,6 @@
 ﻿using Application.Core;
 using Domain.Repositories.Repos.Interfaces;
+using IndentityLogic.Interfaces;
 using MediatR;
 
 namespace Application.Notes
@@ -14,17 +15,26 @@ namespace Application.Notes
         public class Handler : IRequestHandler<Command, Result<MediatR.Unit>>
         {
             private readonly INoteRepository _noteRepository;
+            private readonly IUserAccessor _userAccessor;
 
-            public Handler(INoteRepository noteRepository)
+            public Handler(INoteRepository noteRepository, IUserAccessor userAccessor)
             {
                 _noteRepository = noteRepository;
+                _userAccessor = userAccessor;
             }
 
             public async Task<Result<Unit>> Handle(Command request, CancellationToken cancellationToken)
             {
+                //TODO : when you try to delete unexisted data, you also have "false". Bug or not?)
+
                 var note = await _noteRepository.GetOneAsync(request.Id);
 
                 if (note == null) return null;
+
+                if (!await _noteRepository.IfUserHasAccessToTheNote(request.Id, _userAccessor.GetUserId()))
+                {
+                    return Result<MediatR.Unit>.Failure("You have no access to this data");
+                }
 
                 var result = await _noteRepository.RemoveAsync(note) > 0;
